Fix serious shift-by-one bug in CALIPSO profiles
This is because the vertical extent e3,e4 = 0, -1, but this removes the
last element of the array. The arrays data and height are both
vertically flipped before interpolation, but vertical extent selection
is applied prior, resp. after flipping. This means dimension size is
still corrent, but the arrays are shifted relative to each other by one,
an equivalent of 300m in the worst case.